wip reconfig database
