+++ server/database/database.go
@@ -1,41 +1,35 @@
 package database
 
 import (
-	"database/sql"
 	"log"
 	"os"
 
 	_ "github.com/mattn/go-sqlite3"
 )
 
+// Create enum with type DBType
+type DBType string
+
+const (
+	SQLITE     DBType = "SQLITE"
+	MYSQL      DBType = "MYSQL"
+	POSTGRESQL DBType = "POSTGRESQL"
+	MONGODB    DBType = "MONGODB"
+)
+
 func InitializeSqlLite() {
-	os.Remove("./database/sqlite-database.db")
-	log.Println("Creating sqlite-database.db...")
-	file, err := os.Create("./database/sqlite-database.db") // Create SQLite file
-	if err != nil {
-		log.Fatal(err.Error())
+	log.Println(os.Getenv("DB_TYPE"))
+
+	var db = nil
+
+	switch os.Getenv("DB_TYPE") {
+	case string(MONGODB):
+	case string(MYSQL):
+	case string(POSTGRESQL):
+		log.Print("We do not support this yet...\n using sqlite")
+	case string(SQLITE):
+	default:
+		break
+
 	}
-	file.Close()
-	log.Println("sqlite-database.db created")
-
-	sqliteDatabase, _ := sql.Open("sqlite3", "./database/sqlite-database.db")
-	defer sqliteDatabase.Close()
-	createTable(sqliteDatabase)
-}
-
-func createTable(db *sql.DB) {
-	createUserTableSQL := `CREATE TABLE User (
-		"id" TEXT UNIQUE NOT NULL PRIMARY KEY,		
-		"email" TEXT,
-		"username" TEXT,
-		"isActive" BOOLEAN		
-	  );` // SQL Statement for Create Table
-
-	log.Println("Create user table...")
-	statement, err := db.Prepare(createUserTableSQL) // Prepare SQL Statement
-	if err != nil {
-		log.Fatal(err.Error())
-	}
-	statement.Exec() // Execute SQL Statements
-	log.Println("user table created")
 }
